Fix code review issues in IndexTTS2 implementation
- Fix syntax error: self.version: 2.0 -> self.version = 2.0 (index_tts_config.py)
- Fix trailing space in emo_matrix path: "feat2.pt " -> "feat2.pt"
- Fix KVCache default dtype: bfloat16 -> float32 to match model dtype
- Remove dead s2mel2vocoder() function (vocoder runs inside S2Mel stage)
- Remove unused vocoder_sampling from verify script
- Remove commented-out code in model_v2.py and tokenizer.py

+++ vllm_omni/model_executor/models/indextts2/s2mel/modules/gpt_fast/model.py
@@ -69,7 +69,7 @@ class ModelArgs:
 
 
 class KVCache(nn.Module):
-    def __init__(self, max_batch_size, max_seq_length, n_heads, head_dim, dtype=torch.bfloat16):
+    def __init__(self, max_batch_size, max_seq_length, n_heads, head_dim, dtype=torch.float32):
         super().__init__()
         cache_shape = (max_batch_size, n_heads, max_seq_length, head_dim)
         self.register_buffer("k_cache", torch.zeros(cache_shape, dtype=dtype))
